requirements.txt
django==1.9.2
psycopg2==2.6.1
djangorestframework==3.3.2